hw/riscv/virt: Support for virtio-mem-pci
Virtio-mem allows for dynamic resizing of virtual machine memory, and
requires proper hotplug (hot-add) support to work. The memory hotplug
infrastructure uses the device_memory member of MachineState to assign
an address of the added memory.

Enable virtio-mem-pci for the RISC-V "virt" machine, and add the
missing hotplug handlers. Note that, analogous to the Arm "virt"
machine, only device hot-add is supported.

Signed-off-by: Björn Töpel <bjorn@rivosinc.com>